requirements.txt
PyMuPDF==1.23.8
pdf2image==1.16.3
google-generativeai==0.3.2
python-dotenv==1.0.0
argparse==1.4.0
tqdm==4.66.1
